Our_Agent/Web/requirements.txt
# Web服务器依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# MCP相关依赖
mcp==1.0.0
httpx==0.25.2

# AI API客户端
openai==1.3.7

# 环境变量管理
python-dotenv==1.0.0

# 类型提示
typing-extensions==4.8.0

